// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
// ...
export interface TransporterDetails {
    documents: Array<ExternalBlob>;
    contactPerson: string;
    email: string;
    truckType: TruckType;
    company: string;
    contracts: Array<Contract>;
    address: string;
    phone: string;
    verificationStatus: TransporterVerificationStatus;
}
export interface TrackingUpdate {
    status: string;
    timestamp: bigint;
    location: string;
}
export interface LoadConfirmation {
    orderId: string;
    confirmationFiles: Array<ExternalBlob>;
}
export interface Contract {
    endDate: bigint;
    year: bigint;
    contractText: string;
    startDate: bigint;
}
// ...
export enum ClientVerificationStatus {
    verified = "verified",
    pending = "pending",
    rejected = "rejected"
}
export enum TruckType {
    triaxle = "triaxle",
    superlinkFlatdeck = "superlinkFlatdeck",
    sideTipper = "sideTipper"
}